server/scripts/requirements.txt
accelerate==0.29.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anthropic==0.23.1
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
chromedriver-autoinstaller==0.6.4
click==8.1.7
courlan==1.0.0
cssselect==1.2.0
dataclasses-json==0.6.4
dateparser==1.2.0
decorator==5.1.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
executing==2.0.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.3
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
h11==0.14.0
html2text==2020.1.16
htmldate==1.8.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
ipython==8.23.0
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jusText==3.0.0
langcodes==3.3.0
llama-index==0.10.27
llama-index-agent-openai==0.2.2
llama-index-cli==0.1.11
llama-index-core==0.10.27
llama-index-embeddings-huggingface==0.2.0
llama-index-embeddings-openai==0.1.7
llama-index-indices-managed-llama-cloud==0.1.5
llama-index-legacy==0.9.48
llama-index-llms-anthropic==0.1.9
llama-index-llms-huggingface==0.1.4
llama-index-llms-ollama==0.1.2
llama-index-llms-openai==0.1.14
llama-index-llms-replicate==0.1.3
llama-index-multi-modal-llms-openai==0.1.4
llama-index-program-openai==0.1.5
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.13
llama-index-readers-llama-parse==0.1.4
llama-index-readers-web==0.1.8
llama-index-vector-stores-mongodb==0.1.4
llama-parse==0.4.0
llamaindex-py-client==0.1.16
lxml==5.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
minijinja==1.0.16
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
openai==1.16.2
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.1
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
playwright==1.42.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
pyee==11.0.1
Pygments==2.17.2
pymongo==4.6.3
PyMuPDF==1.24.1
PyMuPDFb==1.24.1
pypdf==4.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
replicate==0.25.1
requests==2.31.0
requests-file==2.0.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
selenium==4.19.0
sentence-transformers==2.6.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
streamlit==1.33.0
striprtf==0.0.26
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tinysegmenter==0.3
tld==0.13
tldextract==5.1.2
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
tornado==6.4
tqdm==4.66.2
trafilatura==1.8.1
traitlets==5.14.2
transformers==4.39.3
trio==0.25.0
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
watchdog==4.0.0
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
